example_url {httr2}	R Documentation

Code for examples

Description

example_url() runs a simple websever using the webfakes package with the following endpoints:

all the ones from the webfakes::httpbin_app()

/iris: paginate through the iris dataset. It has the query parameters page and limit to control the pagination.

example_github_client() is an OAuth client for GitHub.

Usage

example_url(path = "/")

example_github_client()

[Package httr2 version 1.1.2 Index]