httr2-package {httr2}	R Documentation

httr2: Perform HTTP Requests and Process the Responses

Description

Tools for creating and modifying HTTP requests, then performing them and processing the results. 'httr2' is a modern re-imagining of 'httr' that uses a pipe-based interface and solves more of the problems that API wrapping packages face.

Author(s)

Maintainer: Hadley Wickham hadley@posit.co

Other contributors:

Posit Software, PBC [copyright holder, funder]

Maximilian Girlich [contributor]

See Also

Useful links:

https://httr2.r-lib.org

https://github.com/r-lib/httr2

Report bugs at https://github.com/r-lib/httr2/issues

[Package httr2 version 1.1.2 Index]