Create an OAuth app

An OAuth app is the combination of a client, a set of endpoints (i.e. urls where various requests should be sent), and an authentication mechanism. A client consists of at least a client_id, and also often a client_secret. You'll get these values when you create the client on the API's website.

oauth_app(client, endpoints, auth = c("body", "header"))

oauth_client(client_id, client_secret = NULL, name = NULL)

Arguments

client
A OAuth client created with oauth_client().

endpoints
A named character vector of endpoints. The precise endpoints required depend on the flow that you'll use, but all require a token endpoint that returns an access token.

auth
Authentication mechanism used by the API to authenticate confidential requests made during the flow. This is most commonly "body" where the client_id and client_secret are added to the body of the request, but can also be "header" where the client_id and client_secret are sent using the HTTP Authorization header.

client_id
Client identifier.

client_secret
Client secret. This is not technically a confidential pieces of information so you should avoid storing in the public where possible. However, many APIs require it in order to provide a user friendly authentication experience, and the risks of including it are usually low.

name
Optional name for the client. Used when generating cache directory. If NULL, generated from hash of app hostname and client_id.