OAuth client authentication

oauth_client_req_auth() authenticates a request using the authentication strategy defined by the auth and auth_param arguments to oauth_app(). This used to authenticate the client as part of the OAuth flow, not to authenticate a request on behalf of a user.

There are three built-in strategies:

oauth_client_req_body() adds the client id and (optionally) the secret to the request body, as described in rfc6749, Section 2.3.1.

oauth_client_req_header() adds the client id and secret using HTTP basic authentication with the Authorization header, as described in rfc6749, Section 2.3.1.

oauth_client_jwt_rs256() adds a client assertion to the body using a JWT signed with jwt_sign_rs256() using a private key, as described in rfc7523, Section 2.2.

You will generally not call these functions directly but will instead specify them through the auth argument to oauth_client(). The req and client parameters are automatically filled in; other parameters come from the auth_params argument.

oauth_client_req_auth(req, client)

oauth_client_req_auth_header(req, client)

oauth_client_req_auth_body(req, client)

oauth_client_req_auth_jwt_sig(req, client, claim, size = 256, header = list())

Arguments

req
A request.

client
An oauth_client.

claim
Claim set produced by jwt_claim().

size
Size, in bits, of sha2 signature, i.e. 256, 384 or 512. Only for HMAC/RSA, not applicable for ECDSA keys.

header
A named list giving additional fields to include in the JWT header.