oauth_token {httr2}	R Documentation

Create an OAuth token

Description

Creates a S3 object of class <httr2_token> representing an OAuth token returned from the access token endpoint.

Usage

oauth_token(
  access_token,
  token_type = "bearer",
  expires_in = NULL,
  refresh_token = NULL,
  ...,
  .date = Sys.time()
)

Arguments

access_token
The access token used to authenticate request

token_type
Type of token; only "bearer" is currently supported.

expires_in
Number of seconds until token expires.

refresh_token
Optional refresh token; if supplied, this can be used to cheaply get a new access token when this one expires.

...
Additional components returned by the endpoint

.date
Date the request was made; used to convert the relative expires_in to an absolute expires_at.

Value

An OAuth token: an S3 list with class httr2_token.

See Also

oauth_token_cached() to use the token cache with a specified OAuth flow.

Examples

oauth_token("abcdef")
oauth_token("abcdef", expires_in = 3600)
oauth_token("abcdef", refresh_token = "ghijkl")

[Package httr2 version 1.1.2 Index]