Simulate a request

After preparing a request, call req_fetch() to perform it, fetching the results back to R.

req_fetch(req, path = NULL, handle = NULL)

Arguments

req
A request.

path
Optionally, path to save body of request. This is useful for large responses since it avoids storing the response in memory.

handle
Advanced use only; a curl handle.

Value

Returns an HTTP response.

Examples

req("https://google.com") %>%
  req_fetch()

#> <httr2_response>

#> GET https://www.google.com/

#> Status: 200 OK

#> Content-Type: text/html

#> Body: In memory (14950 bytes)