OAuth authentication with a JWT

This uses oauth_flow_jwt() to generate an access token which is then used to authenticate the request with req_auth_bearer_token(). The token is cached in memory.

req_oauth_jwt(
  req,
  app,
  claims,
  signature,
  signature_params = list(),
  scope = NULL,
  token_params = list()
)

Arguments

req
A request.

app
An oauth_app().

claims
A list of claims. If all elements of the claim set are static apart from iat, nbf, exp, or jti, provide a list and jwt_claim_set() will automatically fill in the dynamic components. If other components need to vary, you can instead provide a zero-argument callback function which should call jwt_claim_set().

signature
Function use to sign claim_set, e.g. jwt_sign_rs256().

signature_params
Additional arguments passed to signature

scope
Scopes to be requested from the resource owner.

token_params
List containing additional parameters passed to oauth_flow_access_token().