Define a paginated request

Use req_paginate() to specify how to request the next page in a paginated API. Use req_perform_iteratively() to fetch all pages. If you need more control use a combination of req_perform() and iterate_next_request() to iterate through the pages yourself.

There are also helpers for common pagination patterns:

req_paginate_next_url() when the response contains a link to the next page.

req_paginate_next_token() when the response contains a token that is used to describe the next page.

req_paginate_offset() when the request describes the offset i.e. at which element to start and the page size.

req_paginate_page_index() when the request specifies which the page to request via an index.

Usage

req_paginate(
  req,
  next_request,
  parse_resp = NULL,
  required_fields = NULL,
  n_pages = NULL
)

req_paginate_next_url(req, parse_resp, n_pages = NULL)

req_paginate_token(req, set_token, parse_resp, n_pages = NULL)

req_paginate_offset(req, offset, page_size, parse_resp = NULL, n_pages = NULL)

req_paginate_page_index(req, page_index, parse_resp = NULL, n_pages = NULL)

Arguments

req

A request.

next_request

A callback function that returns a request to the next page or NULL if there is no next page. It takes two arguments:

req: the previous request.

parsed: the previous response parsed via the argument parse_resp.

parse_resp

A function with one argument resp that parses the response and returns a list with the field data and other fields needed to create the request for the next page. req_perform_iteratively() combines all data fields via vctrs::vec_c() and returns the result. Other fields that might be needed are:

* `next_url` for `paginate_next_url()`.
* `next_token` for `paginate_next_token()`.

required_fields

An optional character vector that specifies which fields are required in the list returned by parse_resp().

n_pages

An optional function that extracts the total number of pages, improving the automatically generated progress bar. It has one argument parsed, which is the previous response parsed via the argument parse_resp.

set_token

A function that applies the new token to the request. It takes two arguments: a request and the new token.

req: the previous request.

token: the token for the next page.

offset

A function that applies the new offset to the request. It takes two arguments:

req: the previous request.

offset: the integer offset for the next page.

page_size

A whole number that specifies the page size i.e. the number of elements per page.

page_index

A function that applies the page index to the request. It takes two arguments:

req: the previous request.

offset: the integer page index for the next page.

Value

A modified HTTP request.

See also

Examples

page_size <- 40

request(example_url()) |>
  req_url_path("/iris") |>
  req_url_query(limit = page_size) |>
  req_paginate_page_index(
    page_index = function(req, page) {
      req |> req_url_query(page_index = page)
    },
    parse_resp = function(resp) {
      parsed <- resp_body_json(resp)
      results <- parsed$data
      data <- data.frame(
        Sepal.Length = sapply(results, `[[`, "Sepal.Length"),
        Sepal.Width = sapply(results, `[[`, "Sepal.Width"),
        Petal.Length = sapply(results, `[[`, "Petal.Length"),
        Petal.Width = sapply(results, `[[`, "Petal.Width"),
        Species = sapply(results, `[[`, "Species")
      )

      list(data = data, count = parsed$count)
    },
    n_pages = function(parsed) {
      total <- parsed$count
      ceiling(total / page_size)
    }
  )
#> <httr2_request>
#> GET http://127.0.0.1:35455/iris?limit=40
#> Body: empty
#> Policies:
#> • paginate: a list