Extract the body from the response

resp_body_raw() returns the raw bytes.

resp_body_string() returns a UTF-8 string.

resp_body_json() returns parsed JSON.

resp_body_html() returns parsed HTML.

resp_body_xml() returns parsed XML.

resp_body_json() and resp_body_xml() check that the content-type header is correct; if the server returns an incorrect type you can suppress the check with check_type = FALSE.

resp_body_raw(resp)

resp_body_string(resp, encoding = NULL)

resp_body_json(resp, check_type = TRUE, simplifyVector = FALSE, ...)

resp_body_html(resp, check_type = TRUE, ...)

resp_body_xml(resp, check_type = TRUE, ...)

Arguments

resp
A response object.

encoding
Character encoding of the body text. If not specified, will use the encoding specified by the content-type, falling back to UTF-8 with a warning if it cannot be found. The resulting string is always re-encoded to UTF-8.

check_type
Check that response has expected content type? Set to FALSE to suppress the automated check

simplifyVector
Should JSON arrays containing only primitives (i.e. booleans, numbers, and strings) be caused to atomic vectors?

...
Other argumented passed on to jsonlite::fromJSON() and xml2::read_xml() respectively.