Extract headers from a response

Some of these functions provide generic access to all headers:

resp_headers() retrieves a list of all headers.

resp_header() retrieves a single header.

resp_header_exists() checks if a header is present.

Other functions parse specific headers:

resp_content_type() returns the just the type and subtype of the from the Content-Type header. If Content-Type is not provided (or is not a valid mime type), this returns NA. Used by resp_body_json(), resp_body_html(), and resp_body_xml().

resp_date() returns the Date header as a POSIXct. This header always exists; if the server does not return it, httr2 adds automatically.

resp_encoding() returns the likely character encoding of text types, as parsed from the charset parameter of the Content-Type header. If that header is not found, not valid, or no charset parameter is found, this returns UTF-8. Used by resp_body_string().

resp_retry_after() returns how many seconds you should wait before retrying a request. It parses both forms (absolute and relative) and returns the number of seconds to wait. If the heading is not found, it will return NA.

resp_headers(resp)

resp_header(resp, header)

resp_header_exists(resp, header)

resp_content_type(resp)

resp_date(resp)

resp_encoding(resp)

resp_retry_after(resp)

Arguments

resp
An HTTP response object, as created by req_fetch().

header
Header name (case insensitive)