Temporarily set verbosity for all requests

with_verbosity() is useful for debugging httr2 code buried deep inside another package because it allows you to see exactly what's been sent and requested.

Usage

with_verbosity(code, verbosity = 1)

Arguments

code

Code to execture

verbosity

How much information to print? This is a wrapper around req_verbose() that uses an integer to control verbosity:

0: no output

1: show headers

2: show headers and bodies

3: show headers, bodies, and curl status messages.

Use with_verbosity() to control the verbosity of requests that you can't affect directly.

Value

The result of evaluating code.

Examples

fun <- function() {
  request("https://httr2.r-lib.org") %>% req_perform()
}
with_verbosity(fun())
#> -> GET / HTTP/2
#> -> Host: httr2.r-lib.org
#> -> user-agent: httr2/0.2.3.9000 r-curl/5.0.1 libcurl/7.81.0
#> -> accept: */*
#> -> accept-encoding: deflate, gzip, br, zstd
#> -> 
#> <- HTTP/2 200 
#> <- server: GitHub.com
#> <- content-type: text/html; charset=utf-8
#> <- last-modified: Thu, 22 Jun 2023 18:36:10 GMT
#> <- access-control-allow-origin: *
#> <- etag: W/"6494949a-4cee"
#> <- expires: Sun, 25 Jun 2023 21:33:24 GMT
#> <- cache-control: max-age=600
#> <- content-encoding: gzip
#> <- x-proxy-cache: MISS
#> <- x-github-request-id: 9B38:1848:1046358:186AF58:6498B04C
#> <- accept-ranges: bytes
#> <- date: Sun, 25 Jun 2023 21:23:45 GMT
#> <- via: 1.1 varnish
#> <- age: 21
#> <- x-served-by: cache-iad-kiad7000144-IAD
#> <- x-cache: HIT
#> <- x-cache-hits: 5
#> <- x-timer: S1687728225.344436,VS0,VE1
#> <- vary: Accept-Encoding
#> <- x-fastly-request-id: 010bb710ebc62ac2b47a65534c7190d2025c192d
#> <- content-length: 5312
#> <- 
#> <httr2_response>
#> GET https://httr2.r-lib.org/
#> Status: 200 OK
#> Content-Type: text/html
#> Body: In memory (19694 bytes)