multi_req_perform {httr2}	R Documentation

Perform a list of requests in parallel

Description

Please use req_perform_parallel() instead, and note:

cancel_on_error = FALSE is now on_error = "continue"

cancel_on_error = TRUE is now on_error = "return"

Usage

multi_req_perform(
  reqs,
  paths = NULL,
  pool = deprecated(),
  cancel_on_error = FALSE
)

Arguments

reqs
A list of requests.

paths
An optional character vector of paths, if you want to download the response bodies to disk. If supplied, must be the same length as reqs.

pool
. No longer supported; to control the maximum number of concurrent requests, set max_active.

cancel_on_error
Should all pending requests be cancelled when you hit an error? Set this to TRUE to stop all requests as soon as you hit an error. Responses that were never performed be NULL in the result.

[Package httr2 version 1.1.2 Index]