Automatically cache requests

Use req_perform() to automatically cache HTTP requests. Most API requests are not cacheable, but when possible (often when downloading larger files) it can make a big difference. req_cache() caches responses to GET requests that have status code 200 and at least one of the standard caching headers (e.g. Expires, Etag, Last-Modified, Cache-Control), unless caching has been expressly prohibited with Cache-Control: no-store. Typically, a request will still be sent to the server to check that the cached value is still up-to-date, but it will not need to re-download the body value.

To learn more about HTTP caching, I recommend the MDN article HTTP caching.

req_cache(req, path, use_on_error = FALSE, debug = FALSE)

Arguments

req

A request.

path

Path to cache directory

use_on_error

If the request errors, and there's a cache response, should req_perform() return that instead of generating an error?

debug

When TRUE will emit useful messages telling you about cache hits and misses. This can be helpful to understand whether or not caching is actually doing anything for your use case.