Perform a list of requests in parallel

This variation on req_perform() performs multiple requests in parallel. Unlike req_perform() it always succeeds; it will never throw an error. Instead it will return error objects, which are your responsibility to handle.

Exercise caution when using this function; it's easy to pummel a server with many simultaneous requests. Only use it with hosts designed to serve many files at once.

Usage

req_perform_parallel(reqs, paths = NULL, pool = NULL, cancel_on_error = FALSE)

Arguments

reqs

A list of requests.

paths

An optional list of paths, if you want to download the request bodies to disks. If supplied, must be the same length as reqs.

pool

Optionally, a curl pool made by curl::new_pool(). Supply this if you want to override the defaults for total concurrent connections (100) or concurrent connections per host (6).

cancel_on_error

Should all pending requests be cancelled when you hit an error. Set this to TRUE to stop all requests as soon as you hit an error. Responses that were never performed will have class httr2_cancelled in the result.

Value

A list the same length as reqs where each element is either a response or an error.

Limitations

Will not retrieve a new OAuth token if it expires part way through the requests.

Does not perform throttling with req_throttle().

Does not attempt retries as described by req_retry().

Consults the cache set by req_cache() before/after all requests.

In general, where req_perform() might make multiple requests due to retries or OAuth failures, req_perform_parallel() will only make 1. If any of these limitations are problematic, you may want to use req_perform_sequential() instead.

Examples

# Requesting these 4 pages one at a time would take 2 seconds:
request_base <- request(example_url())
reqs <- list(
  request_base |> req_url_path("/delay/0.5"),
  request_base |> req_url_path("/delay/0.5"),
  request_base |> req_url_path("/delay/0.5"),
  request_base |> req_url_path("/delay/0.5")
)
# But it's much faster if you request in parallel
system.time(resps <- req_perform_parallel(reqs))
#>    user  system elapsed 
#>   0.676   0.361   1.038 

reqs <- list(
  request_base |> req_url_path("/status/200"),
  request_base |> req_url_path("/status/400"),
  request("FAILURE")
)
# req_perform_parallel() will always succeed
resps <- req_perform_parallel(reqs)

# Inspect the successful responses
resps |> resps_successes()
#> [[1]]
#> <httr2_response>
#> GET http://127.0.0.1:44485/status/200
#> Status: 200 OK
#> Content-Type: text/plain
#> Body: None
#> 

# And the failed responses
resps |> resps_failures() |> resps_requests()
#> [[1]]
#> <httr2_request>
#> GET http://127.0.0.1:44485/status/400
#> Body: empty
#> 
#> [[2]]
#> <httr2_request>
#> GET FAILURE
#> Body: empty
#>