Show verbose output when request is performed

req_verbose() uses the following prefixes to distinguish between different components of the HTTP requests and responses:

* informative curl messages

<- request headers

<< request body

-> response headers

>> response body

req_verbose(
  req,
  header_req = TRUE,
  header_resp = TRUE,
  body_req = FALSE,
  body_resp = FALSE,
  info = FALSE,
  redact_header = TRUE
)

Arguments

req
A request.

header_req, header_resp
Show request/response headers?

body_req, body_resp
Should request/response bodies? When the response body is compressed, this will show the number of bytes recevied in each "chunk".

info
Show informational text from curl? This is mainly useful for debugging https and auth problems, so is disabled by default.

redact_header
Redact confidential data in the headers? Currently redacts the contents of the Authorization header to prevent you from accidentally leaking credentials when debugging/reprexing.

See also