Show extra output when request is performed

req_verbose() uses the following prefixes to distinguish between different components of the HTTP requests and responses:

* informative curl messages

<- request headers

<< request body

-> response headers

>> response body

req_verbose(
  req,
  header_req = TRUE,
  header_resp = TRUE,
  body_req = FALSE,
  body_resp = FALSE,
  info = FALSE,
  redact_headers = TRUE
)

Arguments

req

A request.

header_req, header_resp

Show request/response headers?

body_req, body_resp

Should request/response bodies? When the response body is compressed, this will show the number of bytes received in each "chunk".

info

Show informational text from curl? This is mainly useful for debugging https and auth problems, so is disabled by default.

redact_headers

Redact confidential data in the headers? Currently redacts the contents of the Authorization header to prevent you from accidentally leaking credentials when debugging/reprexing.

Value

A modified HTTP request.

See also

Examples

# Use `req_verbose()` to see the headers that are sent back and forth when
# making a request
resp <- request("https://httr2.r-lib.org") %>%
  req_verbose() %>%
  req_perform()
#> -> GET / HTTP/2
#> -> Host: httr2.r-lib.org
#> -> user-agent: httr2/0.1.0.9000 r-curl/4.3.2 libcurl/7.68.0
#> -> accept: */*
#> -> accept-encoding: deflate, gzip, br
#> -> 
#> <- HTTP/2 200 
#> <- server: GitHub.com
#> <- content-type: text/html; charset=utf-8
#> <- last-modified: Thu, 23 Sep 2021 12:33:51 GMT
#> <- access-control-allow-origin: *
#> <- etag: W/"614c742f-4a11"
#> <- expires: Thu, 23 Sep 2021 13:23:17 GMT
#> <- cache-control: max-age=600
#> <- content-encoding: gzip
#> <- x-proxy-cache: MISS
#> <- x-github-request-id: C596:5B7F:754E09:AC8570:614C7D6D
#> <- accept-ranges: bytes
#> <- date: Thu, 23 Sep 2021 13:26:27 GMT
#> <- via: 1.1 varnish
#> <- age: 23
#> <- x-served-by: cache-bwi5033-BWI
#> <- x-cache: HIT
#> <- x-cache-hits: 1
#> <- x-timer: S1632403587.451567,VS0,VE1
#> <- vary: Accept-Encoding
#> <- x-fastly-request-id: e7c74c65e283a18a128e2bc106d481fe63af120a
#> <- content-length: 5316
#> <- 

# Or use one of the convenient shortcuts:
resp <- request("https://httr2.r-lib.org") %>%
  req_perform(verbosity = 1)
#> -> GET / HTTP/2
#> -> Host: httr2.r-lib.org
#> -> user-agent: httr2/0.1.0.9000 r-curl/4.3.2 libcurl/7.68.0
#> -> accept: */*
#> -> accept-encoding: deflate, gzip, br
#> -> 
#> <- HTTP/2 200 
#> <- server: GitHub.com
#> <- content-type: text/html; charset=utf-8
#> <- last-modified: Thu, 23 Sep 2021 12:33:51 GMT
#> <- access-control-allow-origin: *
#> <- etag: W/"614c742f-4a11"
#> <- expires: Thu, 23 Sep 2021 13:23:17 GMT
#> <- cache-control: max-age=600
#> <- content-encoding: gzip
#> <- x-proxy-cache: MISS
#> <- x-github-request-id: C596:5B7F:754E09:AC8570:614C7D6D
#> <- accept-ranges: bytes
#> <- date: Thu, 23 Sep 2021 13:26:27 GMT
#> <- via: 1.1 varnish
#> <- age: 23
#> <- x-served-by: cache-bwi5033-BWI
#> <- x-cache: HIT
#> <- x-cache-hits: 2
#> <- x-timer: S1632403587.468338,VS0,VE0
#> <- vary: Accept-Encoding
#> <- x-fastly-request-id: dbfb93a3b28e588c3e90705c434d02a78e745bfc
#> <- content-length: 5316
#> <-