Show extra output when request is performed

req_verbose() uses the following prefixes to distinguish between different components of the HTTP requests and responses:

* informative curl messages

-> request headers

>> request body

<- response headers

<< response body

Usage

req_verbose(
  req,
  header_req = TRUE,
  header_resp = TRUE,
  body_req = FALSE,
  body_resp = FALSE,
  info = FALSE,
  redact_headers = TRUE
)

Arguments

req

A httr2 request object.

header_req, header_resp

Show request/response headers?

body_req, body_resp

Should request/response bodies? When the response body is compressed, this will show the number of bytes received in each "chunk".

info

Show informational text from curl? This is mainly useful for debugging https and auth problems, so is disabled by default.

redact_headers

Redact confidential data in the headers? Currently redacts the contents of the Authorization header to prevent you from accidentally leaking credentials when debugging/reprexing.

Value

A modified HTTP request.

See also

Examples

# Use `req_verbose()` to see the headers that are sent back and forth when
# making a request
resp <- request("https://httr2.r-lib.org") |>
  req_verbose() |>
  req_perform()
#> -> GET / HTTP/2
#> -> Host: httr2.r-lib.org
#> -> user-agent: httr2/1.0.7.9000 r-curl/6.0.1 libcurl/7.81.0
#> -> accept: */*
#> -> accept-encoding: deflate, gzip, br, zstd
#> -> 
#> <- HTTP/2 200 
#> <- server: GitHub.com
#> <- content-type: text/html; charset=utf-8
#> <- last-modified: Mon, 06 Jan 2025 13:42:09 GMT
#> <- access-control-allow-origin: *
#> <- etag: W/"677bddb1-4983"
#> <- expires: Mon, 06 Jan 2025 13:55:45 GMT
#> <- cache-control: max-age=600
#> <- content-encoding: gzip
#> <- x-proxy-cache: MISS
#> <- x-github-request-id: 20C6:38F320:13636CB:15F7DD2:677BDE89
#> <- accept-ranges: bytes
#> <- date: Mon, 06 Jan 2025 13:46:03 GMT
#> <- via: 1.1 varnish
#> <- age: 18
#> <- x-served-by: cache-iad-kcgs7200107-IAD
#> <- x-cache: HIT
#> <- x-cache-hits: 1
#> <- x-timer: S1736171164.536936,VS0,VE2
#> <- vary: Accept-Encoding
#> <- x-fastly-request-id: 662999d84624414be376182350b7007fe6b4ed62
#> <- content-length: 4638
#> <- 

# Or use one of the convenient shortcuts:
resp <- request("https://httr2.r-lib.org") |>
  req_perform(verbosity = 1)
#> -> GET / HTTP/2
#> -> Host: httr2.r-lib.org
#> -> user-agent: httr2/1.0.7.9000 r-curl/6.0.1 libcurl/7.81.0
#> -> accept: */*
#> -> accept-encoding: deflate, gzip, br, zstd
#> -> 
#> <- HTTP/2 200 
#> <- server: GitHub.com
#> <- content-type: text/html; charset=utf-8
#> <- last-modified: Mon, 06 Jan 2025 13:42:09 GMT
#> <- access-control-allow-origin: *
#> <- etag: W/"677bddb1-4983"
#> <- expires: Mon, 06 Jan 2025 13:55:45 GMT
#> <- cache-control: max-age=600
#> <- content-encoding: gzip
#> <- x-proxy-cache: MISS
#> <- x-github-request-id: 20C6:38F320:13636CB:15F7DD2:677BDE89
#> <- accept-ranges: bytes
#> <- date: Mon, 06 Jan 2025 13:46:03 GMT
#> <- via: 1.1 varnish
#> <- age: 18
#> <- x-served-by: cache-iad-kcgs7200107-IAD
#> <- x-cache: HIT
#> <- x-cache-hits: 2
#> <- x-timer: S1736171164.555882,VS0,VE1
#> <- vary: Accept-Encoding
#> <- x-fastly-request-id: 56a6f599b7431502e7597ea2722c471fa9e8fcc4
#> <- content-length: 4638
#> <-