Extract headers from a response

resp_headers() retrieves a list of all headers.

resp_header() retrieves a single header.

resp_header_exists() checks if a header is present.

Usage

resp_headers(resp, filter = NULL)

resp_header(resp, header)

resp_header_exists(resp, header)

Arguments

resp

An HTTP response object, as created by req_perform().

filter

A regular expression used to filter the header names. NULL, the default, returns all headers.

header

Header name (case insensitive)

Value

resp_headers() returns a list.

resp_header() returns a string if the header exists and NULL otherwise.

resp_header_exists() returns TRUE or FALSE.

Examples

resp <- request("https://httr2.r-lib.org") %>% req_perform()
resp %>% resp_headers()
#> <httr2_headers>
#> server: GitHub.com
#> content-type: text/html; charset=utf-8
#> last-modified: Tue, 03 May 2022 18:45:30 GMT
#> access-control-allow-origin: *
#> etag: W/"6271784a-49b0"
#> expires: Tue, 03 May 2022 19:23:24 GMT
#> cache-control: max-age=600
#> content-encoding: gzip
#> x-proxy-cache: MISS
#> x-github-request-id: 666A:7A20:C332C4:12AFD0B:62717ED4
#> accept-ranges: bytes
#> date: Tue, 03 May 2022 19:13:43 GMT
#> via: 1.1 varnish
#> age: 20
#> x-served-by: cache-dfw18652-DFW
#> x-cache: HIT
#> x-cache-hits: 4
#> x-timer: S1651605224.628857,VS0,VE0
#> vary: Accept-Encoding
#> x-fastly-request-id: 10e9650b6033eb16783cc05324ff9858f0cc5595
#> content-length: 5220
resp %>% resp_headers("x-")
#> <httr2_headers>
#> x-proxy-cache: MISS
#> x-github-request-id: 666A:7A20:C332C4:12AFD0B:62717ED4
#> x-served-by: cache-dfw18652-DFW
#> x-cache: HIT
#> x-cache-hits: 4
#> x-timer: S1651605224.628857,VS0,VE0
#> x-fastly-request-id: 10e9650b6033eb16783cc05324ff9858f0cc5595

resp %>% resp_header_exists("server")
#> [1] TRUE
resp %>% resp_header("server")
#> [1] "GitHub.com"
# Headers are case insensitive
resp %>% resp_header("SERVER")
#> [1] "GitHub.com"

# Returns NULL if header doesn't exist
resp %>% resp_header("this-header-doesnt-exist")
#> NULL