Extract headers from a response

resp_headers() retrieves a list of all headers.

resp_header() retrieves a single header.

resp_header_exists() checks if a header is present.

Usage

resp_headers(resp, filter = NULL)

resp_header(resp, header, default = NULL)

resp_header_exists(resp, header)

Arguments

resp

A httr2 response object, created by req_perform().

filter

A regular expression used to filter the header names. NULL, the default, returns all headers.

header

Header name (case insensitive)

default

Default value to use if header doesn't exist.

Value

resp_headers() returns a list.

resp_header() returns a string if the header exists and NULL otherwise.

resp_header_exists() returns TRUE or FALSE.

Examples

resp <- request("https://httr2.r-lib.org") |> req_perform()
resp |> resp_headers()
#> <httr2_headers>
#> server: GitHub.com
#> content-type: text/html; charset=utf-8
#> last-modified: Wed, 23 Oct 2024 22:11:41 GMT
#> access-control-allow-origin: *
#> etag: W/"6719749d-4983"
#> expires: Thu, 24 Oct 2024 16:14:45 GMT
#> cache-control: max-age=600
#> content-encoding: gzip
#> x-proxy-cache: MISS
#> x-github-request-id: E173:149CD8:117A0FC:1394C8D:671A701D
#> accept-ranges: bytes
#> date: Thu, 24 Oct 2024 16:46:12 GMT
#> via: 1.1 varnish
#> age: 502
#> x-served-by: cache-iad-kcgs7200098-IAD
#> x-cache: HIT
#> x-cache-hits: 4
#> x-timer: S1729788372.471463,VS0,VE0
#> vary: Accept-Encoding
#> x-fastly-request-id: 29da0128760e0f3c748a2e2732e2eafe5be027c6
#> content-length: 4636
resp |> resp_headers("x-")
#> <httr2_headers>
#> x-proxy-cache: MISS
#> x-github-request-id: E173:149CD8:117A0FC:1394C8D:671A701D
#> x-served-by: cache-iad-kcgs7200098-IAD
#> x-cache: HIT
#> x-cache-hits: 4
#> x-timer: S1729788372.471463,VS0,VE0
#> x-fastly-request-id: 29da0128760e0f3c748a2e2732e2eafe5be027c6

resp |> resp_header_exists("server")
#> [1] TRUE
resp |> resp_header("server")
#> [1] "GitHub.com"
# Headers are case insensitive
resp |> resp_header("SERVER")
#> [1] "GitHub.com"

# Returns NULL if header doesn't exist
resp |> resp_header("this-header-doesnt-exist")
#> NULL